refactor: split Table.tsx into usePaymentColumns hook and PaymentTableSummary component
